do not hide magic values
git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@1683 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/netsiddev_builder/NetSIDConnection.java
@@ -191,7 +191,7 @@ public class NetSIDConnection {
 		send(() -> new SetSidLevel(sidNum, volume));
 	}
 
-	public void setBalance(byte sidNum, float balance) {
+	public void setBalance(byte sidNum, byte balance) {
 		send(() -> new SetSidPosition(sidNum, balance));
 	}
